+++ src/main/java/dev/tsvinc/r2p/domain/mapper/R2PTransactionMapper.java
@@ -14,7 +14,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -26,13 +25,13 @@ public class R2PTransactionMapper {
 
     public R2PTransaction createTransactionFromInitiate(InitiateR2pRequest request, PaymentRequestDetail paymentRequest) {
         return R2PTransaction.builder()
-                .paymentRequestId(UUID.randomUUID().toString().substring(0, 21)) // Generate Visa-style ID
+                .paymentRequestId(UUID.randomUUID().toString().substring(0, 21))
                 .endToEndId(paymentRequest.endToEndId())
                 .requestMessageId(request.requestMessageId())
                 .transactionStatus(TransactionStatus.PDNG)
                 .useCase(request.useCase().name())
                 .product(request.product().name())
-                .requestedAmount(BigDecimal.valueOf(paymentRequest.requestedAmount()))
+                .requestedAmount(paymentRequest.requestedAmount()) // Now BigDecimal
                 .requestedAmountCurrency(paymentRequest.requestedAmountCurrency())
                 .creditorAgentId(request.creditor().creditorAgentId())
                 .debtorAgentId(paymentRequest.debtorAgentId())
@@ -42,7 +41,8 @@ public class R2PTransactionMapper {
                 .debtorAlias(paymentRequest.debtorAlias())
                 .debtorAliasType(paymentRequest.debtorAliasType() != null ?
                         paymentRequest.debtorAliasType().name() : null)
-                .dueDate(request.dueDate() != null ? LocalDateTime.parse(request.dueDate() + "T23:59:59").toLocalDate() : null)
+                .dueDate(request.dueDate() != null ?
+                        LocalDateTime.parse(request.dueDate() + "T23:59:59").toLocalDate() : null)
                 .requestReason(serializeToJson(request.requestReason()))
                 .isRefund(false)
                 .createdAt(LocalDateTime.now())
